+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+## Added
+- Introduced multi-format writing for capture files. When format is 'multi' data
+  will be written in both jsonl and parquet formats.
+
 ## [0.29.1]
 ## Fixed
 - Adjust capture manager time syncrhonization to _after_ target is running,